clearCache

+++ android_fw_general/src/main/java/com/xiaocoder/android/fw/general/helper/XCCleanCacheHelper.java
@@ -7,35 +7,34 @@ import com.xiaocoder.android.fw.general.application.XCApp;
 import java.io.File;
 
 /**
- * 删除缓存
+ * 删除缓存 ，removeFileAsyn（文件或文件夹）
  */
 public class XCCleanCacheHelper {
 
-    // public boolean isGoOnDeleting = false;
+    public boolean isGoOnDeleting;
 
     Dialog mDeletingDialog;
 
     public XCCleanCacheHelper(Dialog deletingDialog) {
 
         mDeletingDialog = deletingDialog;
+        // 如果是实际中, 可能存在正在删除, 而页面此时退出的情况
+        isGoOnDeleting = true;
 
     }
 
-    /**
-     * 删除文件夹
-     */
-    public void removeDir(File dir) {
-        // isGoOnDeleting = true; // 如果是实际中, 可能存在正在删除, 而应用此时退出的情况
+    private void removeDir(File dir) {
         if (dir != null && dir.exists() && dir.isDirectory()) {
             File[] files = dir.listFiles();
             if (files != null) {
                 // 如果目录是系统级文件夹，java没有访问权限，那么会返回null数组
                 for (File file : files) {
-                    // if (isGoOnDeleting) {
-                    if (file.isDirectory()) {
-                        removeDir(file);
-                    } else {
-                        file.delete();
+                    if (isGoOnDeleting) {
+                        if (file.isDirectory()) {
+                            removeDir(file);
+                        } else {
+                            file.delete();
+                        }
                     }
                 }
             }
@@ -43,11 +42,10 @@ public class XCCleanCacheHelper {
         }
     }
 
-
+    /**
+     * 子线程中删除
+     */
     public void removeFileAsyn(final File file) {
-        /**
-         * 子线程中删除
-         */
         XCApp.getBase_cache_threadpool().execute(new Runnable() {
             @Override
             public void run() {
@@ -70,22 +68,20 @@ public class XCCleanCacheHelper {
                 } else {
                     file.delete();
                 }
-                XCApp.getBase_handler().post(new Runnable() {
+                XCApp.getBase_handler().postDelayed(new Runnable() {
                     @Override
                     public void run() {
                         if (mDeletingDialog != null) {
-                            mDeletingDialog.dismiss();
+                            mDeletingDialog.cancel();
                         }
                     }
-                });
+                }, 2000);
             }
         });
     }
 
-    //@Override
-    //public void onDestroy() {
-    //    super.onDestroy();
-    //    isGoOnDeleting = false;
-    //}
-
+    public void quit() {
+        isGoOnDeleting = false;
+        mDeletingDialog = null;
+    }
 }
